+++ CHANGELOG.rst
@@ -3,3 +3,5 @@ Changelog
 
 x.x.x (x-x-x)
 -------------------
+
+* Configured redirects for broken links